Add benchmark script.
Benchmarks occur through the application layer by starting a local
ReserveWithUsApp instance and then sending randomly generated queries to
the local instance in a loop. Each query is times and the results are
returned.

For now, the benchmark script may be run in iPython by importing the
benchmark_queries module and the by running the mode benchmark
functions. Future revisions of the script could export to CSV.

This framework may be extended easily to profile additional queries so
long as they are accessible though the ReserveWithUsApp interface.